Functions moved to a separate file. Added 'error.bat' to display errors in release mode
