+++ GuessNumberSFML/GuessNumberSFML.vcxproj.filters
@@ -24,6 +24,9 @@
     <ClCompile Include="..\..\AnimsSFML\AnimsSFML\anims.cpp">
       <Filter>Header Files</Filter>
     </ClCompile>
+    <ClCompile Include="functions.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <Font Include="Res\Bubblegum.ttf">
