+++ GuessNumberSFML/error.bat
@@ -1,0 +1,3 @@
+@echo off
+echo %1 %2 %3
+pause
